feat: dashboard component kit — serve.rs split, print mode, reusable components
Split monolithic serve.rs (8359 lines) into serve/ module directory:
- mod.rs (811) — AppState, router, middleware, utilities
- styles.rs (646) — CSS constant
- js.rs (1193) — GRAPH_JS, SEARCH_JS, AADL_JS
- layout.rs (264) — page_layout + print_layout
- views.rs (5534) — all view handlers
- components.rs (522) — reusable UI kit with 16 unit tests

Component kit: ViewParams, FilterBar, SortableTable, Pagination,
CollapsibleTree — all URL-param-aware for state persistence.

Print mode: ?print=1 strips nav/context bar/HTMX for clean output.
Print button added to context bar.

Also fix pre-commit hooks: +stable for clippy, remove invalid --strict
flag from validate hook.

Implements: FEAT-052
Refs: DD-005
